moving examples and simplifying example to load and display
