--- applications/contrastive_phenotyping/demo/prepare-exercise.sh
@@ -1,5 +1,0 @@
-# Run black on .py files
-# black solution.py
-
-# Convert .py to ipynb
-jupytext --to ipynb --update-metadata '{"jupytext":{"cell_metadata_filter":"all"}}' --update demo_infection.py
