+++ examples/DynaCLR/DynaCLR-DENV-VS-Ph/ipynb_to_py.sh
@@ -1,0 +1,5 @@
+# Run black on .py files
+
+
+# Convert .py to ipynb
+jupytext --to py:percent --update-metadata '{"jupytext":{"cell_metadata_filter":"all"}}' demo_infection.ipynb
