+++ examples/DynaCLR/DynaCLR-DENV-VS-Ph/py_to_ipynb.sh
@@ -1,0 +1,5 @@
+# Run black on .py files
+black solution.py
+
+# Convert .py to ipynb
+jupytext --to ipynb --update-metadata '{"jupytext":{"cell_metadata_filter":"all"}}' --update demo_infection.py
