+++ examples/DynaCLR/README.md
@@ -1,0 +1,7 @@
+# DynaCLR Demos
+
+- [ImageNet vs DynaCLR embeddings (cell infection)](examples/DynaCLR/smooth_embeddings)
+
+
+# Run Inference
+
