+++ examples/configs/dynacell_metrics_test.yml
@@ -1,0 +1,32 @@
+# Example configuration for DynaCell metrics computation
+
+# Required parameters
+target_database: /home/eduardo.hirata/repos/viscy/applications/DynaCell/dynacell_summary_table.csv
+pred_database: /home/eduardo.hirata/repos/viscy/applications/DynaCell/dynacell_summary_table.csv
+output_dir: /home/eduardo.hirata/repos/viscy/applications/DynaCell/metrics_demo_cli
+method: intensity  # Options: 'intensity' or 'segmentation2D'
+
+# Target dataset parameters
+target_channel: Organelle
+target_z_slice: 16  # Use -1 for all slices
+target_cell_types: 
+  - HEK293T
+target_organelles:
+  - HIST2H2BE  
+target_infection_conditions:
+  - Mock
+
+# Prediction dataset parameters
+pred_channel: Organelle
+pred_z_slice: 16
+pred_cell_types:
+  - HEK293T
+pred_organelles: 
+  - HIST2H2BE
+pred_infection_conditions:
+  - Mock
+
+# Processing parameters
+batch_size: 1
+num_workers: 0
+version: "1"
